Heavy refactoring of the news tool to support multiple data loading mechanisms. The default configuration now utilizes an XML file to load news source information.  This may be customized via the newsSpringBeans.xml file in the mdor-webapp project to either load data from a web service or to alternately utilize the previous DB managed news sources.
News related refactoring outside the news tool:
admin tool modified in the publishing controller to point to the new db managed news source. This can be converted later to utilize the spring configuration and the generics.
Config and bootstrapping data files were added to the mdot-webapp project.  A better way to integrate configuration needs to be created so that the tool config can stay with the tool but end up properly installed in the web app.

TODO: 
Code to load news sources from a restful service needs to be completed and tested.
The spring loaded DB driven news sources need to be tested and unit tests written for them.